+++ src/analysis/aperiodicity_extraction_by_world.cc
@@ -47,7 +47,7 @@ bool AperiodicityExtractionByWorld::Run(
   world::InitializeD4COption(&option);
   option.threshold = 0.0;
 
-  const int f0_length(f0.size());
+  const int f0_length(static_cast<int>(f0.size()));
   const double frame_shift_in_sec(frame_shift_ / sampling_rate_);
   std::vector<double> time_axis(f0_length);
   for (int i(0); i < f0_length; ++i) {
